Move all external logic functions from settings to dfLogics
& edit CMakeLists
