+++ TheForceEngine/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(TFE_Archive/)
 add_subdirectory(TFE_Asset/)
 add_subdirectory(TFE_Audio/)
 add_subdirectory(TFE_DarkForces/)
+add_subdirectory(TFE_ExternalData/)
 add_subdirectory(TFE_FileSystem/)
 add_subdirectory(TFE_FrontEndUI/)
 add_subdirectory(TFE_Game/)
