+++ TheForceEngine/TFE_DarkForces/logic.cpp
@@ -479,7 +479,7 @@ namespace TFE_DarkForces
 
 	TFE_ExternalData::CustomActorLogic* tryFindCustomActorLogic(const char* logicName)
 	{
-		TFE_ExternalData::ExternalLogics* externalLogics = TFE_Settings::getExternalLogics();
+		TFE_ExternalData::ExternalLogics* externalLogics = TFE_ExternalData::getExternalLogics();
 		u32 actorCount = externalLogics->actorLogics.size();
 
 		for (u32 a = 0; a < actorCount; a++)
