+++ TheForceEngine/TFE_ExternalData/CMakeLists.txt
@@ -1,0 +1,2 @@
+file(GLOB SOURCES "*.cpp")
+target_sources(tfe PRIVATE ${SOURCES})
