+++ TheForceEngine/TFE_ExternalData/dfLogics.cpp
@@ -2,18 +2,56 @@
 #include <vector>
 #include <TFE_System/cJSON.h>
 #include <TFE_System/system.h>
+#include <TFE_FileSystem/filestream.h>
+#include <TFE_FileSystem/paths.h>
 #include <TFE_DarkForces/time.h>
 #include "dfLogics.h"
 #include "logicTables.h"
 
 namespace TFE_ExternalData
 {
+	static ExternalLogics s_externalLogics = {};
+
 	///////////////////////////
 	/// Forward declarations 
 	///////////////////////////
+	void parseLogicData(char* data, const char* filename, std::vector<CustomActorLogic>& actorLogics);
 	bool tryAssignProperty(cJSON* data, CustomActorLogic& customLogic);
 
 	
+	ExternalLogics* getExternalLogics()
+	{
+		return &s_externalLogics;
+	}
+
+	void loadCustomLogics()
+	{
+		std::vector<string> jsons;
+		TFE_Paths::getAllFilesFromSearchPaths("logics", "json", jsons);
+
+		TFE_System::logWrite(LOG_MSG, "LOGICS", "Parsing logic JSON(s)");
+		for (u32 i = 0; i < jsons.size(); i++)
+		{
+			FileStream file;
+			const char* fileName = jsons[i].c_str();
+			if (!file.open(fileName, FileStream::MODE_READ)) { return; }
+
+			const size_t size = file.getSize();
+			char* data = (char*)malloc(size + 1);
+			if (!data || size == 0)
+			{
+				TFE_System::logWrite(LOG_ERROR, "LOGICS", "JSON found but is %u bytes in size and cannot be read.", size);
+				return;
+			}
+			file.readBuffer(data, (u32)size);
+			data[size] = 0;
+			file.close();
+
+			parseLogicData(data, fileName, s_externalLogics.actorLogics);
+			free(data);
+		}
+	}
+
 	void parseLogicData(char* data, const char* filename, std::vector<CustomActorLogic>& actorLogics)
 	{
 		cJSON* root = cJSON_Parse(data);
